decomposition flashcards api
