+++ src/components/forms/dialog-forms/deck-dialog-form/deck-dialog-form.tsx
@@ -4,8 +4,7 @@ import { useForm } from 'react-hook-form'
 import dummyCover from '@/assets/webp/dummy-cover.webp'
 import { cn } from '@/components/forms/dialog-forms/dialog-forms.styles'
 import { DialogBody as Body, Dialog, DialogContent } from '@/components/ui/primitives'
-import { DeckResponse } from '@/services/decks/deck.types'
-import { useCreateDeckMutation, useUpdateDeckMutation } from '@/services/flashcards-api'
+import { DeckResponse, useCreateDeckMutation, useUpdateDeckMutation } from '@/services'
 import { DIALOG_ACTION } from '@/shared/enums'
 import { deckNameScheme, privateDeckScheme } from '@/shared/schemes'
 import { FlexContainer } from '@/shared/ui/flex-container'
