+++ src/components/ui/layout-components/tables/deck-table/deck-table.stories.tsx
@@ -3,7 +3,7 @@ import type { Meta, StoryObj } from '@storybook/react'
 import { Provider } from 'react-redux'
 import { MemoryRouter } from 'react-router-dom'
 
-import { store } from '@/services/store'
+import { store } from '@/services'
 import { action } from '@storybook/addon-actions'
 
 import { DeckTable as DeckTableComponent } from './deck-table'
