+++ src/pages/deck-page/deck-page.tsx
@@ -3,14 +3,10 @@ import { Link, generatePath, useParams } from 'react-router-dom'
 
 import { ArrowBackOutline } from '@/assets/icons'
 import { CardDialogForm } from '@/components/forms'
-import { DeckTitle } from '@/components/ui/layout-components'
-import { DeckTable } from '@/components/ui/layout-components/tables'
-import { Button, TextField } from '@/components/ui/primitives'
-import { Pagination } from '@/components/ui/primitives/pagination'
+import { DeckTable, DeckTitle } from '@/components/ui/layout-components'
+import { Button, Pagination, TextField } from '@/components/ui/primitives'
 import { cn } from '@/pages/deck-page/deck-page.styles'
-import { PaginationModel } from '@/services/cards/cards.types'
-import { Deck } from '@/services/decks/deck.types'
-import { useGetCardsQuery, useGetDeckQuery } from '@/services/flashcards-api'
+import { Deck, PaginationModel, useGetCardsQuery, useGetDeckQuery } from '@/services'
 import { PATH } from '@/shared/enums'
 import { FlexContainer } from '@/shared/ui/flex-container'
 import { Page } from '@/shared/ui/page'
