+++ src/services/decks/index.ts
@@ -1,0 +1,2 @@
+export * from './deck.types'
+export * from './decks-api'
